install compose via apt
