+++ defaults/main.yml
@@ -1,2 +1,1 @@
 ---
-docker_enabled: no
