+++ molecule/default/molecule.yml
@@ -8,10 +8,10 @@ driver:
 lint:
   name: yamllint
 platforms:
-  - name: stretch
-    box: bento/debian-9.0
-  - name: xenial
-    box: bento/ubuntu-16.04
+- name: stretch
+  box: bento/debian-9
+- name: xenial
+  box: bento/ubuntu-16.04
 provisioner:
   name: ansible
   connection_options:
