+++ molecule/default/playbook.yml
@@ -1,5 +1,6 @@
 ---
 - name: Converge
   hosts: all
+  become: yes
   roles:
     - role: role-docker
