--- molecule/install/.flake8
@@ -1,2 +1,0 @@
-[flake8]
-max-line-length = 130
