--- molecule/install/molecule.yml
@@ -1,26 +1,0 @@
----
-dependency:
-  name: galaxy
-driver:
-  name: vagrant
-  provider:
-    name: virtualbox
-lint:
-  name: yamllint
-platforms:
-  - name: stretch
-    box: bento/debian-9.0
-  - name: xenial
-    box: bento/ubuntu-16.04
-provisioner:
-  name: ansible
-  connection_options:
-    ansible_become: yes
-  lint:
-    name: ansible-lint
-scenario:
-  name: install
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
