+++ requirements.txt
@@ -1,3 +1,3 @@
-ansible
-molecule
-python-vagrant
+ansible==2.6.2
+molecule==2.16.0
+python-vagrant==0.5.15
