--- tasks/docker.yml
@@ -1,17 +1,0 @@
-- name: include distro vars
-  include_vars: '{{ vars_item }}'
-  with_first_found:
-    - "{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml"
-    - "{{ ansible_distribution }}.yml"
-    - "default.yml"
-  loop_control:
-    loop_var: vars_item
-
-- name: include distro installer
-  include_tasks: "{{ distro_installer }}"
-  with_first_found:
-    - "install/{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml"
-    - "install/{{ ansible_distribution }}.yml"
-    - "install/default.yml"
-  loop_control:
-    loop_var: distro_installer
