+++ tasks/install/default.yml
@@ -1,19 +1,19 @@
-- name: install docker deps
+- name: Installing dependencies
   apt:
-    pkg: "{{ pkg_item }}"
+    pkg: "{{ docker_dep_item }}"
     state: present
     update_cache: no
     install_recommends: no
   with_items: '{{ docker_deps }}'
   loop_control:
-    loop_var: pkg_item
+    loop_var: docker_dep_item
 
-- name: add repository key
+- name: Adding repository key
   apt_key:
     url: https://download.docker.com/linux/{{ ansible_distribution.lower() }}/gpg
     id: 0EBFCD88
 
-- name: add repository
+- name: Adding repository
   apt_repository:
     repo: >
       deb [arch=amd64]
@@ -24,20 +24,13 @@
 
 - meta: flush_handlers
 
-- name: install docker
+- name: Installing docker
   apt:
     pkg: docker-ce
-    state: present
     update_cache: no
     install_recommends: yes
   notify: restart docker
 
-# Please note that you need to mount /tmp with exec permissions!
-# -> http://zindo.info/docker-compose-error-while-loading-libz-so-1.html
-- name: install docker-compose
-  get_url:
-    url: https://github.com/docker/compose/releases/download/{{ docker_compose_version }}/docker-compose-{{ ansible_system }}-{{ ansible_architecture }}
-    dest: /usr/local/bin/docker-compose
-    owner: root
-    group: root
-    mode: 0755
+- name: Installing docker-compose
+  package:
+    name: docker-compose
