+++ tasks/main.yml
@@ -1,3 +1,19 @@
----
-- include_tasks: docker.yml
-  when: docker_enabled
+- name: include distro vars
+  include_vars: '{{ vars_item }}'
+  with_first_found:
+    - "{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml"
+    - "{{ ansible_distribution }}.yml"
+    - "default.yml"
+  loop_control:
+    loop_var: vars_item
+  tags: install
+
+- name: include distro installer
+  include_tasks: "{{ distro_installer }}"
+  with_first_found:
+    - "install/{{ ansible_distribution }}.{{ ansible_lsb.codename }}.yml"
+    - "install/{{ ansible_distribution }}.yml"
+    - "install/default.yml"
+  loop_control:
+    loop_var: distro_installer
+  tags: install
